requirements/test.txt
# Test dependencies go here.
-r base.txt

black==19.3b0
coverage==4.5.3
coveralls==1.8.0
django-test-plus==1.2.0
factory-boy==2.12.0
fakeredis==1.0.3
flake8==3.7.7
pytest-django==3.5.0
pytest-factoryboy==2.0.2
pytest-mock==1.10.4
pytest-tap==2.3
responses==0.10.6
